Remove indirection around barcode callouts
